홈화면 진행중인 교환 수정
- ExchangeProgressAdapter → TrackerAdapter로 교체
- 교환 타입별(택배/직접/함께) 네비게이션 분기 처리
- ExchangeRole을 TrackerData에 추가해 탭 기반 → role 기반으로 전환
- toTrackerData() 로직을 TrackerDataMapper.kt로 분리
- DirectGuest/DirectHostActivity WindowInsets 처리 추가
- .gitattributes 추가 (줄바꿈 LF 통일)
