+++ .gitattributes
@@ -1,0 +1,23 @@
+# 모든 텍스트 파일을 LF로 통일
+* text=auto eol=lf
+
+# Windows에서도 LF 유지
+*.kt text eol=lf
+*.kts text eol=lf
+*.java text eol=lf
+*.xml text eol=lf
+*.json text eol=lf
+*.md text eol=lf
+*.html text eol=lf
+*.gradle text eol=lf
+*.properties text eol=lf
+*.pro text eol=lf
+
+# 바이너리 파일은 변환 안 함
+*.png binary
+*.jpg binary
+*.jpeg binary
+*.gif binary
+*.webp binary
+*.keystore binary
+*.jks binary
